Added remove all edited objects button

+++ src/reducers/objects-edit.js
@@ -137,15 +137,22 @@ function resetEditedObjects(state, action) {
 }
 
 
-/*
-    Removes edited objects with provided `objectIDs` from the state. 
+/**
+    Removes edited objects from the state.
+
+    If `removeAll` flag is true, removes all edited objects.
+    
+    Otherwise, removes objects with provided `objectIDs`.
     Removes all new subobjects of composite objects.
     If `removeSubobjects` is true, also removes existing subobjects of composite objects.
 */
 function removeEditedObjects(state, action) {
-    const { objectIDs, removeSubobjects } = action;
+    const { objectIDs, removeSubobjects, removeAll } = action;
 
-    // Remove edited objects and their non-composite children
+    // Remove all edited objects
+    if (removeAll) return { ...state, editedObjects: {} };
+
+    // Remove edited objects specified in `objectIDs` and their non-composite children
     return getStateWithRemovedEditedObjects(state, objectIDs, { deleteAllSubobjects: removeSubobjects });
 }
